requirements.txt
fastapi
uvicorn
pydantic
google-genai
yfinance
ccxt
vnstock
python-dotenv
requests
pandas
langgraph
langchain-google-genai
langchain-core
